Added Compile instruction and CI integration with style checker and javadoc
This commit adds a compile instruction to know which steps are necessary on local build.
For the continouse integration a travis script is needed.
This will instruct travis to build HELIOS according the maven script
and publish a binary version of the release on tagging.

The maven build description was extended to run checkstyle on compile.
Additional targets are the API documentation and binary package preparing.