Fixed GPS time calculation and integrated it to the scanner
The calculation of the GPS time had a calculation offset error. I hop i fix it.
Additional a added a parameter to the doSimStep of the scanner. This parameter allows to have one time base for the hole simulation. The simulation now don't use the System Time as timebase.

For verification I extended the singleRayPulse and measurement by the gpsTime.