Refactored start of LidarSimulation Application
The decoding of commandline arguments was mixed with initialization
of the simulator. The refactoring separated this steps. This allow
later to reuse the LidarSim class or unit test it without configuration
by commandline arguments syntax.

The class LidarSim handles the setup of the simulation for visualization
or for processing.
The new class CliParser is parsing the arguments and converts them into
information which are independent from the decoding. This class is only
used in the main for commandline argument parsing.